PDF-Sorter/requirements.txt
streamlit
langchain-groq
python-dotenv
pypdf